+++ nemo/utils/model_utils.py
@@ -47,11 +47,6 @@ except ModuleNotFoundError:
 MODEL_CONFIG = "model_config.yaml"
 _VAL_TEST_FASTPATH_KEY = 'ds_item'
 
-# NeMo2 checkpoint structure is a checkpoint directory, with a WEIGHTS_PATH and CONTEXT_PATH subdirectory structure.
-#  WEIGHTS_PATH stores the weights while CONTEXT_PATH stores the hyper-parameters.
-WEIGHTS_PATH: str = "weights"
-CONTEXT_PATH: str = "context"
-
 
 class ArtifactPathType(Enum):
     """
@@ -682,8 +677,9 @@ def ckpt_to_dir(filepath: Union[str, Path]) -> Path:
     """
 
     filepath = Path(filepath)
-    if not filepath.suffix == ".ckpt":
-        filepath = filepath.with_suffix(filepath.suffix + ".ckpt")
+    # if it is already a distributed checkpoint, then return
+    if filepath.suffix != ".ckpt" and filepath.is_dir():
+        return filepath
 
     # adding this assert because we will later remove directories based on the return value of this method
     assert filepath.suffix == ".ckpt", f"filepath: {filepath} must have .ckpt extension"
@@ -692,16 +688,6 @@ def ckpt_to_dir(filepath: Union[str, Path]) -> Path:
     checkpoint_dir = filepath.with_name(filepath.stem)
 
     return checkpoint_dir
-
-
-def ckpt_to_weights_subdir(filepath: Union[str, Path]) -> Path:
-    """Given an input checkpoint filepath, clean it using `ckpt_to_dir` and then return the weights subdirectory."""
-    return ckpt_to_dir(filepath=filepath) / WEIGHTS_PATH
-
-
-def ckpt_to_context_subdir(filepath: Union[str, Path]) -> Path:
-    """Given an input checkpoint filepath, clean it using `ckpt_to_dir` and then return the context subdirectory."""
-    return ckpt_to_dir(filepath=filepath) / CONTEXT_PATH
 
 
 def save_artifacts(model, output_dir: str, use_abspath: bool = False) -> None:
